spilitfile
